% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parsers.R
\name{is_plotly}
\alias{is_plotly}
\title{Determine if a plot is Plotly}
\usage{
is_plotly(plot_data)
}
\arguments{
\item{plot_data}{A list containing the top level data for a single plot.}
}
\value{
A logical scalar. TRUE if the plot is a plotly plot, or FALSE if it's a HighCharts one.
}
\description{
Plotly plots are generated by MultiQC 1.2 and above.
This matters because the plot data formats differ between these two formats.
}
\keyword{internal}
