% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parsers.R
\name{map_xy_line_datasets}
\alias{map_xy_line_datasets}
\title{Parses a list of xyline plot datasets}
\usage{
map_xy_line_datasets(datasets, plot_name)
}
\arguments{
\item{dataset}{A list which has the names "name" and "data", which
originate from a MultiQC plot}
}
\value{
A list whose names are sample names. Each value is a data frame with one column.
}
\description{
Parses a list of xyline plot datasets
}
\keyword{internal}
