% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\docType{package}
\name{MultiqcAnalyse}
\alias{MultiqcAnalyse}
\title{MultiqcAnalyse: A package for computating the notorious bar statistic}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

