% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parse_bar_graph}
\alias{parse_bar_graph}
\title{Takes the JSON dictionary for a bar graph, and returns a list of lists
of quality metrics}
\usage{
parse_bar_graph(plot_data, prefix, extractor, summary)
}
\description{
Takes the JSON dictionary for a bar graph, and returns a list of lists
of quality metrics
}
\keyword{internal}
