% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{parse_general}
\alias{parse_general}
\title{Parses the "general_stats_data_ section}
\usage{
parse_general(parsed)
}
\arguments{
\item{parsed}{The full parsed multiqc JSON file}
}
\value{
A list of samples, each of which has a list of metrics
}
\description{
Parses the "general_stats_data_ section
}
\keyword{internal}
