% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{parse_metadata}
\alias{parse_metadata}
\title{Parses metadata using a user-supplied function}
\usage{
parse_metadata(parsed, samples, find_metadata)
}
\arguments{
\item{parsed}{The parsed multiqc.json}

\item{samples}{A list of known sample names}
}
\description{
Parses metadata using a user-supplied function
}
\keyword{internal}
