% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parse_plot_features}
\alias{parse_plot_features}
\title{Returns a list of summary statistics for a plotly plot, provided as a list
e.g. from jsonlite}
\usage{
parse_plot_features(
  plot_data,
  prefix,
  extractor = extract_ignore_x,
  summary = list(mean = mean)
)
}
\arguments{
\item{plot_data}{A list containing the keys $plot_type, $datasets and $config}

\item{prefix}{The prefix for this plot type in the final data frame}

\item{extractor}{A function which converts the raw plot JSON into a vector}

\item{summary}{A function that maps a vector to a scalar}
}
\value{
A list of samples, each containing a list of plots, each containing
a list of summary stats
}
\description{
Returns a list of summary statistics for a plotly plot, provided as a list
e.g. from jsonlite
}
