% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parse_plots}
\alias{parse_plots}
\title{Parses the "report_saved_raw_data" section}
\usage{
parse_plots(parsed, options)
}
\arguments{
\item{parsed}{The full parsed multiqc JSON file}
}
\value{
A list of samples, each of which has a list of metrics
}
\description{
Parses the "report_saved_raw_data" section
}
\keyword{internal}
