% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parse_xyline_plot}
\alias{parse_xyline_plot}
\title{Takes the JSON dictionary for an xyline plot, and returns a list of lists
of quality metrics}
\usage{
parse_xyline_plot(plot_data, prefix, extractor, summary = list(mean = mean))
}
\description{
Takes the JSON dictionary for an xyline plot, and returns a list of lists
of quality metrics
}
\keyword{internal}
