% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_features}
\alias{plot_features}
\title{Returns a list of summary statistics for the provided plot data}
\usage{
plot_features(...)
}
\arguments{
\item{plot_data}{A list containing the keys $plot_type, $datasets and $config}

\item{type}{The name of this plot, e.g. "fastqc_per_base_n_content_plot"}

\item{extractor}{A function which converts the raw plot JSON into a vector}

\item{summary}{A function that maps a vector to a scalar}

\item{rename}{A new name for this plot}
}
\value{
A list of samples, each containing a list of plots, each containing
a list of summary stats
}
\description{
Returns a list of summary statistics for the provided plot data
}
\keyword{internal}
