% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_to_histogram}
\alias{plot_to_histogram}
\title{Given a plot list column for a histogram type plot,
converts into a list of HistDat objects,
which can then be used to efficiently calculate summary statistics}
\usage{
plot_to_histogram(list_col)
}
\arguments{
\item{list_col}{A list containing tibbles}
}
\value{
A list of HistDat objects
}
\description{
Given a plot list column for a histogram type plot,
converts into a list of HistDat objects,
which can then be used to efficiently calculate summary statistics
}
