% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sanitise_column_name}
\alias{sanitise_column_name}
\title{Converts a MultiQC string into a standardised identifier for the final
data frame}
\usage{
sanitise_column_name(name)
}
\description{
Converts a MultiQC string into a standardised identifier for the final
data frame
}
\keyword{internal}
