% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stat_q30}
\alias{stat_q30}
\title{Summary statistic for finding the Q30 of a dataset of quality scores}
\usage{
stat_q30(vec)
}
\description{
Summary statistic for finding the Q30 of a dataset of quality scores
}
