% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_q30}
\alias{summary_q30}
\title{Summary statistic for finding the \%Q30 of a dataset of quality scores
This is the proportion of total reads in a dataset that have a quality
score of 30 or above.}
\usage{
summary_q30(vec)
}
\arguments{
\item{vec}{Either a \link[HistDat:HistDat-class]{HistDat::HistDat} or a 1-D numeric vector}
}
\value{
The \%Q30 of the dataset, as a numeric of length 1
}
\description{
Summary statistic for finding the \%Q30 of a dataset of quality scores
This is the proportion of total reads in a dataset that have a quality
score of 30 or above.
}
