% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\docType{package}
\name{TidyMultiqc-package}
\alias{TidyMultiqc-package}
\title{TidyMultiqc: Converting MultiQC reports into tidy data frames}
\description{
This package provides the means to convert \code{multiqc_data.json}
files, produced by the wonderful \href{http://multiqc.info/}{MultiQC} tool,
into tidy data.frames for downstream analysis in R.
If you are reading this, you should immediately stop reading this and
go to https://github.com/multimeric/TidyMultiqc
}
