% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{extract_histogram}
\alias{extract_histogram}
\title{Extractor function that calculates statistics for a histogram. For example
this might be relevant for the}
\usage{
extract_histogram(data, as_hist_dat = T)
}
\description{
Extractor function that calculates statistics for a histogram. For example
this might be relevant for the
}
