% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{load_multiqc}
\alias{load_multiqc}
\title{Loads a collection of MultiQC JSON files}
\usage{
load_multiqc(paths, ...)
}
\arguments{
\item{paths}{A vector of paths to MultiQC JSON files}

\item{...}{Args to pass to load_multiqc_file}
}
\value{
A tibble, see load_multiqc_file
}
\description{
Loads a collection of MultiQC JSON files
}
