% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parsers.R
\name{parse_bar_graph}
\alias{parse_bar_graph}
\title{Takes the JSON dictionary for a bar graph, and returns a named list of
data frames, one for each sample.}
\usage{
parse_bar_graph(plot_data, name)
}
\value{
A list of data frames, one for each sample.
Each data frame will have one column corresponding to each category in the bar chart.
For example, for the plot "SnpEff: Counts by Genomic Region", we will have
one column for the number of intron variants, one column for the number of exon variants, etc.
This means that the number of columns will be fairly variable for different plots.
}
\description{
Takes the JSON dictionary for a bar graph, and returns a named list of
data frames, one for each sample.
}
\keyword{internal}
\keyword{plot_parser}
