% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\docType{package}
\name{TidyMultiqc-package}
\alias{TidyMultiqc}
\alias{TidyMultiqc-package}
\title{TidyMultiqc: Converting MultiQC reports into tidy data frames}
\description{
This package provides the means to convert \code{multiqc_data.json}
files, produced by the wonderful \href{http://multiqc.info/}{MultiQC} tool,
into tidy data.frames for downstream analysis in R.
If you are reading this manual, you should immediately stop reading this and
instead refer to the documentation website at \url{https://multimeric.github.io/TidyMultiqc/}, which
provides more accessible documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://multimeric.github.io/TidyMultiqc/}
  \item \url{https://github.com/multimeric/TidyMultiqc}
  \item \url{https://cran.r-project.org/package=TidyMultiqc}
  \item Report bugs at \url{https://github.com/multimeric/TidyMultiqc/issues}
}

}
